Виправлено алгоритм Рабіна-Міллера

+++ Lab2/lab2_functions.hxx
@@ -6,7 +6,7 @@ VeryLongInt* karatsuba(VeryLongInt x, VeryLongInt y);
 VeryLongInt* toomKuk(VeryLongInt x, VeryLongInt y);
 VeryLongInt* shenhage(VeryLongInt x, VeryLongInt y);
 VeryLongInt* shtrassen(VeryLongInt x, VeryLongInt y);
-bool millerRabin(unsigned long long x, int k);
-bool solovayStrassen(unsigned long long x, int k);
+bool millerRabin(long x, long k);
+bool solovayStrassen(long x, int k);
 
 #endif // LAB2_FUNCTIONS_HXX_INCLUDED
